Move credential files to oauth2 subtree.
