+++ google/cloud/storage/client_options.h
@@ -16,7 +16,7 @@
 #define GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_STORAGE_CLIENT_OPTIONS_H_
 
 #include "google/cloud/internal/throw_delegate.h"
-#include "google/cloud/storage/credentials.h"
+#include "google/cloud/storage/oauth2/credentials.h"
 
 namespace google {
 namespace cloud {
