+++ google/cloud/storage/oauth2/authorized_user_credentials.h
@@ -15,8 +15,8 @@
 #ifndef GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_STORAGE_INTERNAL_AUTHORIZED_USER_CREDENTIALS_H_
 #define GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_STORAGE_INTERNAL_AUTHORIZED_USER_CREDENTIALS_H_
 
-#include "google/cloud/storage/credentials.h"
-#include "google/cloud/storage/internal/credential_constants.h"
+#include "credentials.h"
+#include "credential_constants.h"
 #include "google/cloud/storage/internal/curl_request_builder.h"
 #include "google/cloud/storage/internal/nljson.h"
 #include <chrono>
