+++ google/cloud/storage/oauth2/credential_constants.h
