+++ google/cloud/storage/oauth2/credentials.h
