+++ google/cloud/storage/oauth2/google_application_default_credentials_file.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "google/cloud/storage/internal/google_application_default_credentials_file.h"
+#include "google_application_default_credentials_file.h"
 #include "google/cloud/internal/throw_delegate.h"
 #include <sstream>
