+++ google/cloud/storage/oauth2/service_account_credentials_test.cc
@@ -12,9 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "google/cloud/storage/internal/service_account_credentials.h"
+#include "service_account_credentials.h"
 #include "google/cloud/internal/setenv.h"
-#include "google/cloud/storage/internal/credential_constants.h"
+#include "credential_constants.h"
 #include "google/cloud/storage/internal/nljson.h"
 #include "google/cloud/storage/testing/mock_http_request.h"
 #include <gmock/gmock.h>
