+++ google/cloud/storage/storage_client.bzl
@@ -6,13 +6,11 @@ storage_client_HDRS = [
     "client_options.h",
     "credentials.h",
     "internal/access_control_common.h",
-    "internal/authorized_user_credentials.h",
     "internal/binary_data_as_debug_string.h",
     "internal/bucket_acl_requests.h",
     "internal/bucket_requests.h",
     "internal/raw_client_wrapper_utils.h",
     "internal/common_metadata.h",
-    "internal/credential_constants.h",
     "internal/curl_handle.h",
     "internal/curl_handle_factory.h",
     "internal/curl_download_request.h",
@@ -27,7 +25,6 @@ storage_client_HDRS = [
     "internal/format_rfc3339.h",
     "internal/generic_object_request.h",
     "internal/generic_request.h",
-    "internal/google_application_default_credentials_file.h",
     "internal/http_response.h",
     "internal/logging_client.h",
     "internal/metadata_parser.h",
@@ -41,7 +38,6 @@ storage_client_HDRS = [
     "internal/patch_builder.h",
     "internal/raw_client.h",
     "internal/retry_client.h",
-    "internal/service_account_credentials.h",
     "internal/service_account_requests.h",
     "lifecycle_rule.h",
     "list_buckets_reader.h",
@@ -49,6 +45,10 @@ storage_client_HDRS = [
     "notification_event_type.h",
     "notification_metadata.h",
     "notification_payload_format.h",
+    "oauth2/authorized_user_credentials.h",
+    "oauth2/credential_constants.h",
+    "oauth2/google_application_default_credentials_file.h",
+    "oauth2/service_account_credentials.h",
     "object_access_control.h",
     "object_metadata.h",
     "object_rewriter.h",
@@ -84,7 +84,6 @@ storage_client_SRCS = [
     "internal/default_object_acl_requests.cc",
     "internal/empty_response.cc",
     "internal/format_rfc3339.cc",
-    "internal/google_application_default_credentials_file.cc",
     "internal/http_response.cc",
     "internal/logging_client.cc",
     "internal/metadata_parser.cc",
@@ -100,6 +99,7 @@ storage_client_SRCS = [
     "list_buckets_reader.cc",
     "list_objects_reader.cc",
     "notification_metadata.cc",
+    "oauth2/google_application_default_credentials_file.cc",
     "object_access_control.cc",
     "object_metadata.cc",
     "object_rewriter.cc",
