+++ google/cloud/storage/storage_client_unit_tests.bzl
@@ -12,13 +12,11 @@ storage_client_unit_tests = [
     "client_write_object_test.cc",
     "credentials_test.cc",
     "internal/access_control_common_test.cc",
-    "internal/authorized_user_credentials_test.cc",
     "internal/binary_data_as_debug_string_test.cc",
     "internal/bucket_acl_requests_test.cc",
     "internal/bucket_requests_test.cc",
     "internal/default_object_acl_requests_test.cc",
     "internal/format_rfc3339_test.cc",
-    "internal/google_application_default_credentials_file_test.cc",
     "internal/http_response_test.cc",
     "internal/logging_client_test.cc",
     "internal/metadata_parser_test.cc",
@@ -29,11 +27,13 @@ storage_client_unit_tests = [
     "internal/parse_rfc3339_test.cc",
     "internal/patch_builder_test.cc",
     "internal/retry_client_test.cc",
-    "internal/service_account_credentials_test.cc",
     "internal/service_account_requests_test.cc",
     "lifecycle_rule_test.cc",
     "list_buckets_reader_test.cc",
     "list_objects_reader_test.cc",
+    "oauth2/authorized_user_credentials_test.cc",
+    "oauth2/google_application_default_credentials_file_test.cc",
+    "oauth2/service_account_credentials_test.cc",
     "object_access_control_test.cc",
     "object_metadata_test.cc",
     "object_test.cc",
